+++ CharityHub.Api/CharityHub.Domain/Entities/User.cs
@@ -14,7 +14,7 @@ namespace CharityHub.Domain.Entities
         public virtual Sex Sex { get; set; }
 
 
-        public virtual long? CharityId { get; set; }
+        public virtual int? CharityId { get; set; }
         public virtual Charity Charity { get; set; }
 
         public virtual ICollection<User_Charity> ObservedCharities { get; set; }
