+++ CharityHub.Api/CharityHub.Domain/Entities/User_Charity.cs
@@ -2,10 +2,10 @@
 {
     public class User_Charity : Entity
     {
-        public virtual long UserId { get; set; }
+        public virtual int UserId { get; set; }
         public virtual User User { get; set; }
 
-        public virtual long CharityId { get; set; }
+        public virtual int CharityId { get; set; }
         public virtual Charity Charity { get; set; }
     }
 }
